attempting to update just pandas
travis doesn't seem to be able to update all of conda. I tried
running it without running the
"sudo conda update conda --yes" line. This failed with a pandas
error message stating that there is no module named xlrd. The API
for pandas recently changed, and we need to use the latest version
for this code to work. Therefore, I need to update pandas.

Rather than updating the entirety of conda, can we just update
pandas? This commit attempts to do just that. For starters, I
did NOT include `sudo`. I have a feeling this will fail. If it does
I will include `sudo` as the next step.